refactor(commands): extract module-find/lazy-import/watch/profile into commands/tooling.rs (Issue #186)
Continues the PR-A10 incremental decomposition of src/commands/mod.rs by
moving the pybun module-find, lazy-import, watch, and profile command
implementations into a new src/commands/tooling.rs module. Pure code
move, no behavior change. mod.rs shrinks from 5957 to 5241 lines.
