+++ docs/PLAN.md
@@ -167,7 +167,9 @@
   - Goal: `src/commands.rs`（6841行）を thin な `execute()` dispatcher として残し、各コマンドの実装をドメイン別モジュールへ段階的に抽出する。挙動変更なし、純粋なコード移動であることを既存テストスイートで保証する。
   - Current: `src/commands.rs` を `src/commands/mod.rs` へ `git mv`。テスト実行系（discovery/native executor/snapshot正規化/compat診断/`run_tests` 本体、約1076行）を `src/commands/test.rs` へ抽出し、`mod test;` を宣言。`mod.rs` 側の呼び出しを `test::run_tests(...)` に更新。`RenderDetail`（コンストラクタ含む）と `find_python_interpreter()` は `mod.rs` に private のまま残し、Rust の子モジュールから祖先モジュールの private 項目へアクセス可能というルールにより `super::` 経由で参照（追加の `pub(crate)` 化は不要）。`#[cfg(test)] mod tests` 内の重複していたテスト関数群（shard parsing / snapshot normalization / compat diagnostic、8件）を `mod.rs` から削除し `commands/test.rs` の `#[cfg(test)] mod tests` へ移動。`mod.rs` は 7043行 → 5855行に縮小。
   - Tests: `cargo build` / `cargo clippy --all-targets --all-features -- -D warnings` / `cargo fmt -- --check` / `cargo test --lib`（351 passed）/ `cargo test --lib commands::test::`（8/8 passed）/ `cargo test`（フルスイート、E2E含む全テストパス）。JSON/CLI の公開挙動・schema に変更なし。
-  - Remaining: `install.rs` / `lock.rs` / `run.rs` / `build.rs` / `x.rs` / `maintenance.rs`（gc/doctor）/ `python_env.rs` / `tooling.rs`（module-find/lazy-import/watch/profile）/ `project.rs`（init/outdated/upgrade）への抽出は follow-up PR で対応する。
+  - Current: `pybun module-find` / `pybun lazy-import` / `pybun watch` / `pybun profile` の実装（`run_module_find` / `run_lazy_import` / `run_watch` / `run_profile`、約715行、関連する `module_finder` / `hot_reload` / `lazy_import` / `profiles` の局所 `use` を含む）を `src/commands/tooling.rs` へ抽出し、`mod tooling;` を宣言。`mod.rs` 側の呼び出しを `tooling::run_module_find(...)` 等に更新し、`ModuleFindArgs` / `LazyImportArgs` / `WatchArgs` / `ProfileArgs` の import を `mod.rs` から削除して `tooling.rs` 側に移動。`RenderDetail`（コンストラクタ含む）は `super::RenderDetail` 経由で参照。`mod.rs` は 5957行 → 5241行に縮小。
+  - Tests: `cargo build` / `cargo clippy --all-targets --all-features -- -D warnings` / `cargo fmt -- --check` / `cargo test --lib`（354 passed）/ `cargo test`（フルスイート、E2E含む全テストパス）。JSON/CLI の公開挙動・schema に変更なし。
+  - Remaining: `install.rs` / `lock.rs` / `run.rs` / `build.rs` / `x.rs` / `maintenance.rs`（gc/doctor）/ `python_env.rs` / `project.rs`（init/outdated/upgrade）への抽出は follow-up PR で対応する。
 
 ### P2 (Post-GA Improvement)
 - PR-A8: Runtime catalog hardening（実チェックサム管理 + 3.13 preview）
